# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjetTutore API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/listings": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Liste des annonces",
                "operationId": "0c014ef6c84a38a711a9dfa90af19ed5",
                "parameters": [
                    {
                        "name": "currency_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_amount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exchange_rate",
                                "amount_available",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des annonces"
                    }
                }
            },
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Publier une nouvelle annonce",
                "operationId": "e1c7a71d62affda164d97989f5903699",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "currency_from",
                                    "currency_to",
                                    "amount_available",
                                    "exchange_rate"
                                ],
                                "properties": {
                                    "currency_from": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "currency_to": {
                                        "type": "string",
                                        "example": "EUR"
                                    },
                                    "amount_available": {
                                        "type": "number",
                                        "example": 100.5
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 0.92
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Annonce créée avec succès"
                    },
                    "403": {
                        "description": "KYC requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/listings/{id}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Afficher une annonce",
                "operationId": "1f5d43b628ab16b04a86b5494a0255ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'annonce"
                    },
                    "404": {
                        "description": "Annonce introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Listings"
                ],
                "summary": "Modifier une annonce",
                "operationId": "2271e0ff5b518140f02106073fc8d1cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "currency_from": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "currency_to": {
                                        "type": "string",
                                        "example": "EUR"
                                    },
                                    "amount_available": {
                                        "type": "number",
                                        "example": 150
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 0.95
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Annonce mise à jour"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Listings"
                ],
                "summary": "Supprimer une annonce",
                "operationId": "d9557f5899a494fc1bd64cbb2afb5a81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Annonce supprimée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Transactions en cours"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}